data-lander-py/requirements.txt
boto3
python-dotenv
requests
pandas
pyarrow